fixes #48 close the open chanel before waiting for nextReader
When sending Binary data, the socket.io client strips the binary
data from the frame and replaces it with a numbered _placeholder.
Once the placeholder frame is ack'ed the client sends another
frame with the binary data.  Without this fix go-socket.io never
acks the placeholder frame and so waits indefinitely for the data
frame.